[BC-5] Refactored sql native criteria builder package. Readme has been improved as well
