+++ src/main/java/com/bc/ecommerce/application/usescases/PricesUseCase.java
@@ -1,10 +1,8 @@
 package com.bc.ecommerce.application.usescases;
 
-import com.bc.ecommerce.domain.business.sql.QueryBuilder;
 import com.bc.ecommerce.domain.operational.Prices;
 import com.bc.ecommerce.domain.port.in.PricesService;
 import com.bc.ecommerce.domain.port.out.PricesRepository;
-import com.bc.ecommerce.infrastructure.db.springdata.query.CustomQuery;
 import com.bc.ecommerce.infrastructure.rest.spring.pojo.PricesCriteria;
 import lombok.AllArgsConstructor;
 
@@ -25,8 +23,7 @@ public class PricesUseCase implements PricesService {
      */
     @Override
     public Prices search(PricesCriteria criteria) {
-        CustomQuery query  = QueryBuilder.retrieve(criteria);
-        return repository.pricesProjection(query);
+        return repository.pricesProjection(criteria);
     }
 
 }
